Flask/requirements.txt
flask==2.3.3
flask-cors==3.0.10
tensorflow==2.18.0
pillow==9.5.0
numpy==1.26.0

